+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0] - 2024-05-09
+
 ### Added
 
 - File types for Svelte, Nuxt config (`nuxt.config.ts`), NPM package (`package.json`, `package-lock.json`), TypeScript config (`tsconfig.json`)
@@ -91,7 +93,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release
 
-[Unreleased]: https://github.com/pandier/intellij-discord-rp/compare/v1.2.1...HEAD
+[Unreleased]: https://github.com/pandier/intellij-discord-rp/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/pandier/intellij-discord-rp/compare/v1.2.1...v1.3.0
 [1.2.1]: https://github.com/pandier/intellij-discord-rp/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/pandier/intellij-discord-rp/compare/v1.1.1...v1.2.0
 [1.1.1]: https://github.com/pandier/intellij-discord-rp/compare/v1.1.0...v1.1.1